Add files via upload
Removed jquery parsehtml as a vulnerability in sanitizing text in chat/issuer-profiles (thanks Markus Alvila for notifying on this).  
Disabled chatbox entirely out of precaution and to eliminate user input on the site unless the user explicitly browses to an issuer's token.  
Those who still want to see and experiment with the chatbox can re-enable it by setting var chatEnabled to true in the js file.